requirements.txt
nextcord~=2.0.0
python-dotenv~=0.20.0